+++ tests/test_cli.py
@@ -2,7 +2,6 @@ import textwrap
 from unittest.mock import patch
 
 from click.testing import CliRunner
-
 from hibpcli.cli import main
 from hibpcli.exceptions import ApiError
 from hibpcli.password import Password
@@ -69,7 +68,7 @@ def test_keepass_subcommand_with_path_and_password_options(mock_check):
 @patch.object(Password, "is_leaked", return_value=True)
 def test_password_subcommand_for_leaked_password(mock_password):
     runner = CliRunner()
-    result = runner.invoke(main, ["password", "--password", "test"])
+    result = runner.invoke(main, ["check-password", "--password", "test"])
     expected_output = "Please change your password!\n"
     assert result.output == expected_output
 
@@ -77,7 +76,7 @@ def test_password_subcommand_for_leaked_password(mock_password):
 @patch.object(Password, "is_leaked", return_value=False)
 def test_password_subcommand_for_safe_password(mock_password):
     runner = CliRunner()
-    result = runner.invoke(main, ["password", "--password", "test"])
+    result = runner.invoke(main, ["check-password", "--password", "test"])
     expected_output = "Your password is safe!\n"
     assert result.output == expected_output
 
@@ -85,7 +84,7 @@ def test_password_subcommand_for_safe_password(mock_password):
 @patch.object(Password, "is_leaked", return_value=False)
 def test_password_subcommand_with_prompt(mock_password):
     runner = CliRunner()
-    result = runner.invoke(main, ["password"], input="test")
+    result = runner.invoke(main, ["check-password"], input="test")
     expected_output = """\
         Please enter a password which should be checked: 
         Your password is safe!
@@ -107,6 +106,6 @@ def test_keepass_subcommand_error_handling(mock_password):
 @patch.object(Password, "is_leaked", side_effect=ApiError("Error"))
 def test_password_subcommand_error_handling(mock_password):
     runner = CliRunner()
-    result = runner.invoke(main, ["password", "--password", "test"])
+    result = runner.invoke(main, ["check-password", "--password", "test"])
     expected_output = "Error\n"
     assert result.output == expected_output