+++ CHANGES.rst
@@ -14,6 +14,7 @@ unreleased
 - use setup.py instead of flit for packaging
 - move source code in src directory
 - update versions of installation dependencies
+- subcommand to check a single password now is name `check-password`
 
 
 0.3.0 (03.11.2019)